requirements.txt
python-dotenv
fastapi
uvicorn
openai
google-api-python-client
apscheduler
